sphere-backend/requirement.txt
fastapi
uvicorn
requests
python-dotenv
supabase
jwt
python-jose
pydantic
